@@ -2,7 +2,7 @@
 name = "inventory-app"
 version = "2.0.0"
 description = "Flask Inventory Management Application"
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
     "flask>=3.0.0",
     "flask-sqlalchemy>=3.1.0",
