@@ -1,133 +1,67 @@
+"""Flask inventory management application."""
+
+from __future__ import annotations
+
+import csv
+import io
+import json
 import os
 import sys
-import io
-import csv
-import json
+from typing import TYPE_CHECKING, Any
+
 from flask import Flask, render_template, request
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import func
 
-inventoryApp = Flask(__name__)
-inventoryApp.config.from_object(
-    os.environ.get("APP_SETTINGS", "config.DevelopmentConfig")
-)
-inventoryApp.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-db = SQLAlchemy(inventoryApp)
+if TYPE_CHECKING:
+    from sqlalchemy.orm import Query
+    from werkzeug.datastructures import FileStorage
 
+app = Flask(__name__)
+app.config.from_object(os.environ.get("APP_SETTINGS", "config.DevelopmentConfig"))
+app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+db = SQLAlchemy(app)
+
+# Import models after db is created to avoid circular import
 from models import Grocery  # noqa: E402
 
 
-@inventoryApp.route("/", methods=["GET", "POST"])
-def index():
-    errors = []
-    items = []
-    col = "ID"
-    load_search = False
-    load_add_item = True
-    load_add_csv = False
-    item_searched = False
-    item_added = False
-    if request.method == "POST":
-        if "search-item" in request.form:
-            load_search = True
-            load_add_item = False
-            load_add_csv = False
-            return render_template(
-                "index.html",
-                errors=errors,
-                items=items,
-                column=col,
-                loading_search=load_search,
-                loading_add_item=load_add_item,
-                loading_add_csv=load_add_csv,
-                item_searched=item_searched,
-                item_added=item_added,
-            )
-        elif "add-item" in request.form:
-            load_search = False
-            load_add_item = True
-            load_add_csv = False
-            return render_template(
-                "index.html",
-                errors=errors,
-                items=items,
-                column=col,
-                loading_search=load_search,
-                loading_add_item=load_add_item,
-                loading_add_csv=load_add_csv,
-                item_searched=item_searched,
-                item_added=item_added,
-            )
-        elif "add-csv" in request.form:
-            load_search = False
-            load_add_item = False
-            load_add_csv = True
-            return render_template(
-                "index.html",
-                errors=errors,
-                items=items,
-                column=col,
-                loading_search=load_search,
-                loading_add_item=load_add_item,
-                loading_add_csv=load_add_csv,
-                item_searched=item_searched,
-                item_added=item_added,
-            )
-        elif "send-search" in request.form:
-            load_search = True
-            load_add_item = False
-            load_add_csv = False
-            item_searched = True
-            item_added = False
-            try:
-                col = request.form["column"]
-                item = request.form["item"]
-                matching_items = get_matching_items(col, item)
-                for item in matching_items:
-                    items.append(item)
-            except Exception as ex:
-                error_type = "Unable to search for item. Please double check your search parameters. "
-                errors = report_exception(ex, error_type, errors)
-        elif "send-add" in request.form:
-            load_search = False
-            load_add_item = True
-            load_add_csv = False
-            item_searched = False
-            try:
-                item = set_item()
-                errors, items = add_item(item, errors, items)
-                item_added = True
-            except Exception as ex:
-                error_type = (
-                    "Unable to add item. Please double check your item parameters. "
-                )
-                errors = report_exception(ex, error_type, errors)
-        elif "csv-submit" in request.form:
-            load_search = False
-            load_add_item = False
-            load_add_csv = True
-            item_searched = False
-            try:
-                file = request.files["csv-input"]
-                if not file:
-                    errors.append("No file")
-                if "." in file.filename:
-                    file_ext = file.filename.rsplit(".", 1)[1].lower()
-                else:
-                    errors.append("Invalid file type. Needs to be .csv")
-                if file_ext == "csv":
-                    stream = io.StringIO(
-                        file.stream.read().decode("UTF8"), newline=None
-                    )
-                    csv_input = csv.reader(stream)
-                    iterate_through_csv(csv_input, errors, items)
-                else:
-                    errors.append("Invalid file type. Needs to be .csv")
-            except Exception as ex:
-                error_type = (
-                    "Unable to add item. Please double check your item parameters. "
-                )
-                errors = report_exception(ex, error_type, errors)
+class FormAction:
+    """Constants for form actions."""
+
+    SEARCH_ITEM = "search-item"
+    ADD_ITEM = "add-item"
+    ADD_CSV = "add-csv"
+    SEND_SEARCH = "send-search"
+    SEND_ADD = "send-add"
+    CSV_SUBMIT = "csv-submit"
+
+
+def render_index_template(
+    errors: list[str],
+    items: list[Any],
+    col: str,
+    load_search: bool,
+    load_add_item: bool,
+    load_add_csv: bool,
+    item_searched: bool,
+    item_added: bool,
+) -> str:
+    """Render the index template with given parameters.
+
+    Args:
+        errors: List of error messages.
+        items: List of items to display.
+        col: Column name for search.
+        load_search: Whether to load the search form.
+        load_add_item: Whether to load the add item form.
+        load_add_csv: Whether to load the CSV upload form.
+        item_searched: Whether a search was performed.
+        item_added: Whether an item was added.
+
+    Returns:
+        Rendered HTML template.
+    """
     return render_template(
         "index.html",
         errors=errors,
@@ -141,63 +75,192 @@ def index():
     )
 
 
-def report_exception(ex, error_type, errors):
-    error_str = (
-        error_type
-        + str(ex)
-        + " - Error on line no: {}".format(sys.exc_info()[-1].tb_lineno)
+def handle_search_action() -> tuple[list[str], list[Any], bool, bool]:
+    """Handle search form submission.
+
+    Returns:
+        Tuple of (errors, items, item_searched, item_added).
+    """
+    errors: list[str] = []
+    items: list[Any] = []
+    try:
+        col = request.form["column"]
+        search_item = request.form["item"]
+        matching_items = get_matching_items(col, search_item)
+        items.extend(matching_items)
+    except (KeyError, ValueError, TypeError) as ex:
+        error_type = "Unable to search for item. Please double check your search parameters. "
+        errors = report_exception(ex, error_type, errors)
+    return errors, items, True, False
+
+
+def handle_add_action() -> tuple[list[str], list[Any], bool, bool]:
+    """Handle add item form submission.
+
+    Returns:
+        Tuple of (errors, items, item_searched, item_added).
+    """
+    errors: list[str] = []
+    items: list[Any] = []
+    try:
+        item = set_item()
+        errors, items = add_item(item, errors, items)
+        item_added = True
+    except (KeyError, ValueError, TypeError) as ex:
+        error_type = "Unable to add item. Please double check your item parameters. "
+        errors = report_exception(ex, error_type, errors)
+        item_added = False
+    return errors, items, False, item_added
+
+
+def handle_csv_action() -> tuple[list[str], list[Any]]:
+    """Handle CSV upload form submission.
+
+    Returns:
+        Tuple of (errors, items).
+    """
+    errors: list[str] = []
+    items: list[Any] = []
+    try:
+        file: FileStorage | None = request.files.get("csv-input")
+        if not file or not file.filename:
+            errors.append("No file")
+            return errors, items
+
+        if "." not in file.filename:
+            errors.append("Invalid file type. Needs to be .csv")
+            return errors, items
+
+        file_ext = file.filename.rsplit(".", 1)[1].lower()
+        if file_ext == "csv":
+            stream = io.StringIO(file.stream.read().decode("UTF8"), newline=None)
+            csv_input = csv.reader(stream)
+            iterate_through_csv(csv_input, errors, items)
+        else:
+            errors.append("Invalid file type. Needs to be .csv")
+    except (KeyError, ValueError, TypeError, UnicodeDecodeError) as ex:
+        error_type = "Unable to process CSV file. Please check the file format. "
+        errors = report_exception(ex, error_type, errors)
+    return errors, items
+
+
+@app.route("/", methods=["GET", "POST"])
+def index() -> str:
+    """Handle the main index page for inventory management.
+
+    Returns:
+        Rendered HTML template.
+    """
+    errors: list[str] = []
+    items: list[Any] = []
+    col = "ID"
+    load_search = False
+    load_add_item = True
+    load_add_csv = False
+    item_searched = False
+    item_added = False
+
+    if request.method == "POST":
+        # Handle form view switching
+        if FormAction.SEARCH_ITEM in request.form:
+            load_search, load_add_item, load_add_csv = True, False, False
+        elif FormAction.ADD_ITEM in request.form:
+            load_search, load_add_item, load_add_csv = False, True, False
+        elif FormAction.ADD_CSV in request.form:
+            load_search, load_add_item, load_add_csv = False, False, True
+        # Handle form submissions
+        elif FormAction.SEND_SEARCH in request.form:
+            load_search, load_add_item, load_add_csv = True, False, False
+            errors, items, item_searched, item_added = handle_search_action()
+        elif FormAction.SEND_ADD in request.form:
+            load_search, load_add_item, load_add_csv = False, True, False
+            errors, items, item_searched, item_added = handle_add_action()
+        elif FormAction.CSV_SUBMIT in request.form:
+            load_search, load_add_item, load_add_csv = False, False, True
+            errors, items = handle_csv_action()
+
+    return render_index_template(
+        errors, items, col, load_search, load_add_item, load_add_csv, item_searched, item_added
     )
+
+
+def report_exception(ex: Exception, error_type: str, errors: list[str]) -> list[str]:
+    """Report an exception by adding it to the errors list.
+
+    Args:
+        ex: The exception that occurred.
+        error_type: Description of the error type.
+        errors: Current list of errors.
+
+    Returns:
+        Updated list of errors.
+    """
+    exc_tb = sys.exc_info()[-1]
+    tb_lineno: int | str = exc_tb.tb_lineno if exc_tb is not None else "unknown"
+    error_str = f"{error_type}{ex!s} - Error on line no: {tb_lineno}"
     errors.append(error_str)
     for error in errors:
-        print(error)
+        print(error)  # noqa: T201
     return errors
 
 
-def get_matching_items(search_column, search_item):
+def get_matching_items(search_column: str, search_item: str) -> Query[Any] | dict[str, Any]:
+    """Get items matching the search criteria.
+
+    Args:
+        search_column: Column to search in.
+        search_item: Value to search for.
+
+    Returns:
+        Query result with matching items or empty dict.
+    """
+    # Basic SQL injection protection
     if "DROP TABLE" in search_item:
         return {}
-    if search_column == "id":
-        if search_item.isdigit():
-            return Grocery.query.filter(Grocery.id == int(search_item))
-        else:
+
+    # Handle exact integer searches for id and x_for columns
+    if search_column in ("id", "x_for"):
+        if not search_item.isdigit():
             return {}
-    elif search_column == "x_for":
-        if search_item.isdigit():
-            return Grocery.query.filter(Grocery.x_for == int(search_item))
-        else:
-            return {}
-    elif "*" in search_item or "_" in search_item:
+        column = Grocery.id if search_column == "id" else Grocery.x_for
+        return Grocery.query.filter(column == int(search_item))  # type: ignore[no-any-return]
+
+    # Build search term based on input
+    if "*" in search_item or "_" in search_item:
         search_term = search_item.replace("_", "__").replace("*", "%").replace("?", "_")
-    elif search_item[-1] == "s":
-        search_term = search_item[0 : len(search_item) - 1]
-        search_term = "%{0}%".format(search_term)
+    elif search_item and search_item[-1] == "s":
+        search_term = f"%{search_item[:-1]}%"
     else:
-        search_term = "%{0}%".format(search_item)
+        search_term = f"%{search_item}%"
+
+    # Build and return query
     if search_column == "last_sold":
-        return Grocery.query.filter(
-            func.to_char(Grocery.last_sold, "%YYYY-MM-DD%").ilike(
-                search_term,
-            )
-        ).order_by(Grocery.id)
-    return Grocery.query.filter(
-        getattr(Grocery, search_column).ilike(
-            search_term,
+        query = Grocery.query.filter(
+            func.to_char(Grocery.last_sold, "%YYYY-MM-DD%").ilike(search_term)
         )
-    ).order_by(Grocery.id)
+    else:
+        query = Grocery.query.filter(getattr(Grocery, search_column).ilike(search_term))
+
+    return query.order_by(Grocery.id)  # type: ignore[no-any-return]
 
 
-def set_item():
-    id = int(request.form["id-add"])
+def set_item() -> Grocery:
+    """Create a Grocery item from form data.
+
+    Returns:
+        New Grocery instance.
+    """
+    item_id = int(request.form["id-add"])
     description = request.form["description-add"]
     last_sold = request.form["last-sold-add"]
     shelf_life = request.form["shelf-life-add"]
     department = request.form["department-add"]
     price = request.form["price-add"]
     unit = request.form["unit-add"]
-    x_for = request.form["xfor-add"]
+    x_for = int(request.form["xfor-add"])
     cost = request.form["cost-add"]
-    item_to_add = Grocery(
-        id=id,
+    return Grocery(
+        item_id=item_id,
         description=description,
         last_sold=last_sold,
         shelf_life=shelf_life,
@@ -207,45 +270,57 @@ def set_item():
         x_for=x_for,
         cost=cost,
     )
-    return item_to_add
 
 
-def add_item(item, errors, items):
+def add_item(item: Grocery, errors: list[str], items: list[Any]) -> tuple[list[str], list[Any]]:
+    """Add a grocery item to the database.
+
+    Args:
+        item: Grocery item to add.
+        errors: Current list of errors.
+        items: Current list of items.
+
+    Returns:
+        Tuple of (errors, items).
+    """
     try:
         # SQLAlchemy 2.0 compatible exists check
-        item_exists = (
-            db.session.query(Grocery).filter(Grocery.id == item.id).first() is not None
-        )
+        item_exists = db.session.query(Grocery).filter(Grocery.id == item.id).first() is not None
         if not item_exists:
             db.session.add(item)
             db.session.commit()
             json_obj = json.dumps(dict(item))
             items.append(json_obj)
         else:
-            errors.append(
-                "Unable to add item to database. This item has already been added with ID: "
-                + str(item.id)
-            )
-    except Exception as ex:
+            errors.append(f"Unable to add item to database. This item has already been added with ID: {item.id}")
+    except (ValueError, TypeError) as ex:
         db.session.rollback()
-        errors.append("Unable to add item to database. " + str(ex))
+        errors.append(f"Unable to add item to database. {ex!s}")
     return errors, items
 
 
-def iterate_through_csv(input, errors, items):
-    count = 0
-    for row in input:
-        if count != 0:
+def iterate_through_csv(
+    csv_input: Any, errors: list[str], items: list[Any]
+) -> None:
+    """Process CSV input and add items to database.
+
+    Args:
+        csv_input: CSV reader object.
+        errors: Current list of errors.
+        items: Current list of items.
+    """
+    row: list[str]
+    for idx, row in enumerate(csv_input):
+        if idx != 0:  # Skip header row
             csv_item_to_add = Grocery(
-                id=row[0],
+                item_id=int(row[0]),
                 description=row[1],
                 last_sold=row[2],
                 shelf_life=row[3],
                 department=row[4],
                 price=row[5],
                 unit=row[6],
-                x_for=row[7],
+                x_for=int(row[7]),
                 cost=row[8],
             )
             add_item(csv_item_to_add, errors, items)
-        count += 1
