@@ -1,13 +1,14 @@
 import os
-from flask_script import Manager
+
 from flask_migrate import Migrate, MigrateCommand
+from flask_script import Manager
 
-from inventoryApp import inventoryApp, db
+from inventoryApp import app, db
 
-inventoryApp.config.from_object(os.environ["APP_SETTINGS"])
+app.config.from_object(os.environ["APP_SETTINGS"])
 
-migrate = Migrate(inventoryApp, db)
-manager = Manager(inventoryApp)
+migrate = Migrate(app, db)
+manager = Manager(app)
 
 manager.add_command("db", MigrateCommand)
 
